Implement featuer of  load data from Open Trivia DB
This implementation is not meant to be used on the UI sections, this is a lower level API.
This should turn into a highe-level abstraction to be used by the UI and deliver the best experience to the end user with realibility.

This code also uses axio because this is required by the challenge, but this could be easily replaced by `fetch`, which works out-of-the box.